+++ .prettierrc.js
@@ -1,3 +1,4 @@
+// TODO: is this doing anything at all?
 module.exports = {
   printWidth: 70,
 };