+++ screens/Library.js
@@ -30,9 +30,6 @@ function Library(navigation, route) {
   useEffect(() => {
     async function getStories() {
       try {
-        // const expenses = await fetchExpenses();
-        // load from mock
-        // const allStories = [...data.stories];
         const allStories = await fetchStories();
         console.log("loaded stories");
         setStories(allStories);
@@ -61,6 +58,7 @@ function Library(navigation, route) {
         // console.log("Setting settings in globalConfig");
         setGlobalConfig(JSON.parse(settings));
       }
+      return settings;
     }
     setIsFetching(true);
     getSettings();